requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-core==0.1.18
langchain-openai==0.0.5
langchain-anthropic==0.0.3
langchain-groq==0.0.1
langgraph==0.0.40
langgraph-checkpoint==0.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
uuid==1.30
typing-extensions==4.8.0
tiktoken==0.5.2